requirements.txt
asgiref==3.3.1
boto3==1.16.56
botocore==1.19.56
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.4
django-allauth==0.44.0
django-autoslug==1.9.8
django-countries==7.0
django-crispy-forms==1.10.0
django-js-asset==1.2.2
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
oauthlib==3.1.0
Pillow==8.0.1
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
